refactor: 불필요한 의존성 주입 및 주석 제거
